+++ src/everos/memory/search/__init__.py
@@ -22,12 +22,14 @@ External usage::
         SearchAgentSkillItem,
         SearchAtomicFactItem,
         SearchData,
+        SearchDecisionItem,
         SearchEpisodeItem,
         SearchMethod,
         SearchProfileItem,
         SearchRequest,
         SearchResponse,
         compile_filters,
+        compile_filters_for_decision,
         compile_predicate,
     )
 
@@ -41,6 +43,7 @@ from .dto import SearchAgentCaseItem as SearchAgentCaseItem
 from .dto import SearchAgentSkillItem as SearchAgentSkillItem
 from .dto import SearchAtomicFactItem as SearchAtomicFactItem
 from .dto import SearchData as SearchData
+from .dto import SearchDecisionItem as SearchDecisionItem
 from .dto import SearchEpisodeItem as SearchEpisodeItem
 from .dto import SearchMethod as SearchMethod
 from .dto import SearchProfileItem as SearchProfileItem
@@ -50,6 +53,7 @@ from .filters import ALLOWED_FIELDS as ALLOWED_FIELDS
 from .filters import RESERVED_FIELDS as RESERVED_FIELDS
 from .filters import FilterError as FilterError
 from .filters import compile_filters as compile_filters
+from .filters import compile_filters_for_decision as compile_filters_for_decision
 from .filters import compile_predicate as compile_predicate
 
 __all__ = [
@@ -61,11 +65,13 @@ __all__ = [
     "SearchAgentSkillItem",
     "SearchAtomicFactItem",
     "SearchData",
+    "SearchDecisionItem",
     "SearchEpisodeItem",
     "SearchMethod",
     "SearchProfileItem",
     "SearchRequest",
     "SearchResponse",
     "compile_filters",
+    "compile_filters_for_decision",
     "compile_predicate",
 ]